trouble shooting re-render
