+++ package.json
@@ -12,7 +12,7 @@
     "eslint": "eslint '**/*.{js,jsx}'",
     "eslint:fix": "npm run eslint -- --fix",
     "prepare": "npx husky install",
-    "stylelint": "npx stylelint src/{**/*,*}.{css,js,jsx}",
+    "stylelint": "npx stylelint 'src/{**/*,*}.{css,js,jsx}'",
     "storybook": "start-storybook -p 6006 -s public",
     "storybook:build": "build-storybook -s public",
     "test:unit": "npm run jest",
