+++ src/pages/questions.js
@@ -1,4 +1,3 @@
-import {nanoid} from 'nanoid';
 import {Helmet} from 'react-helmet';
 
 import Layout from '../components/Layout';
@@ -7,21 +6,6 @@ import {useStore} from '../hooks/useStore';
 
 export default function QuestionPage() {
 	const questions = useStore(state => state.questions);
-	const addNewAnswer = useStore(state => state.questions);
-
-	const addAnswer = (event, id) => {
-		event.preventDefault();
-		const index = questions.findIndex(question => {
-			return id === question.id;
-		});
-		const newAnswer = {
-			id: nanoid(),
-			answerText: event.target.inputField.value,
-		};
-		const newQuestions = [...questions];
-		newQuestions[index].answers = [...newQuestions[index].answers, newAnswer];
-		addNewAnswer(newQuestions);
-	};
 
 	return (
 		<Layout>
@@ -35,13 +19,7 @@ export default function QuestionPage() {
 			</Helmet>
 			<article>
 				{questions.map(question => {
-					return (
-						<QuestionComponent
-							key={question.id}
-							question={question}
-							addAnswer={addAnswer}
-						/>
-					);
+					return <QuestionComponent key={question.id} question={question} />;
 				})}
 			</article>
 		</Layout>
